mySearchEngine/requirements.txt
argcomplete==1.9.2
asgiref==3.5.0
base58==0.2.4
boto3==1.21.24
botocore==1.24.24
certifi==2021.10.8
cfn-flip==1.0.4
charset-normalizer==2.0.12
click==8.0.4
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
docutils==0.18.1
durationpy==0.5
future==0.16.0
hjson==3.0.1
httpie==3.1.0
idna==3.3
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
multidict==6.0.2
placebo==0.9.0
Pygments==2.11.2
PySocks==1.7.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2021.3
PyYAML==3.12
requests==2.27.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.19.1
troposphere==3.2.2
Unidecode==1.3.4
urllib3==1.26.8
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.54.1
